+++ test-web/src/components/DataListView/Children/ExcelImportAlertModal.js
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import { Backdrop, Box, Modal, Fade, Button, Typography } from '@mui/material';
+import style from '../style/excelimportalertmodalstyle';
 
 const navy = '#0F3659';
 
@@ -78,16 +79,3 @@ const ExcelImportAlertModal = ({
 };
 
 export default ExcelImportAlertModal;
-
-const style = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 500,
-  bgcolor: 'background.paper',
-  //border: '2px solid #000',
-  boxShadow: 24,
-  p: 4,
-  borderRadius: '10px',
-};
