+++ test-web/src/components/DataListView/style/excelimportalertmodalstyle.js
@@ -1,0 +1,14 @@
+const style = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: 500,
+    bgcolor: 'background.paper',
+    //border: '2px solid #000',
+    boxShadow: 24,
+    p: 4,
+    borderRadius: '10px',
+  };
+
+export default style
